refactor: 패키지 레이어 재구성
