+++ src/main/java/back/ecommerce/domain/product/Product.java
@@ -7,7 +7,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-import back.ecommerce.domain.BaseTimeEntity;
+import back.ecommerce.domain.common.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
