+++ src/main/java/back/ecommerce/dto/request/cart/AddCartRequest.java
@@ -1,7 +1,9 @@
-package back.ecommerce.dto.request;
+package back.ecommerce.dto.request.cart;
 
 import javax.validation.constraints.Email;
+import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -10,10 +12,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class LoginRequest {
+public class AddCartRequest {
 	@NotBlank(message = "이메일은 필수적으로 필요합니다.")
 	@Email(message = "옳바른 이메일 형식이 아닙니다.")
 	private String email;
-	@NotBlank(message = "비밀번호는 필수적으로 필요합니다.")
-	private String password;
+	@NotNull(message = "상품의 아이디는 필수적으로 필요합니다.")
+	private Long productId;
+	@Min(value = 1,message = "상품의 갯수는 최소 1개 이상이여야 합니다.")
+	private Integer quantity;
+
 }
