+++ src/main/java/back/ecommerce/dto/response/common/SuccessResponse.java
@@ -1,4 +1,4 @@
-package back.ecommerce.dto.response;
+package back.ecommerce.dto.response.common;
 
 import lombok.Getter;
