+++ src/main/java/back/ecommerce/dto/response/product/ProductDto.java
@@ -1,4 +1,4 @@
-package back.ecommerce.dto;
+package back.ecommerce.dto.response.product;
 
 import back.ecommerce.domain.product.Category;
 import lombok.AllArgsConstructor;
