+++ src/main/java/back/ecommerce/repository/product/ProductRepository.java
@@ -1,4 +1,4 @@
-package back.ecommerce.repository;
+package back.ecommerce.repository.product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
