+++ src/main/java/back/ecommerce/repository/user/UserRepository.java
@@ -1,10 +1,10 @@
-package back.ecommerce.repository;
+package back.ecommerce.repository.user;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import back.ecommerce.domain.User;
+import back.ecommerce.domain.user.User;
 
 public interface UserRepository extends JpaRepository<User,Long> {
 	Optional<User> findByEmail(String email);
