+++ src/main/java/back/ecommerce/service/auth/AuthService.java
@@ -1,4 +1,4 @@
-package back.ecommerce.service;
+package back.ecommerce.service.auth;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -6,11 +6,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import back.ecommerce.auth.token.Token;
 import back.ecommerce.auth.token.TokenProvider;
-import back.ecommerce.domain.User;
-import back.ecommerce.dto.response.token.TokenResponseDto;
+import back.ecommerce.domain.user.User;
+import back.ecommerce.dto.response.auth.TokenResponseDto;
 import back.ecommerce.exception.PasswordNotMatchedException;
 import back.ecommerce.exception.UserNotFoundException;
-import back.ecommerce.repository.UserRepository;
+import back.ecommerce.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
