+++ src/main/java/back/ecommerce/service/product/ProductService.java
@@ -1,4 +1,4 @@
-package back.ecommerce.service;
+package back.ecommerce.service.product;
 
 import static back.ecommerce.constant.PageConstant.*;
 
@@ -8,10 +8,10 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import back.ecommerce.domain.product.Category;
-import back.ecommerce.dto.ProductDto;
-import back.ecommerce.dto.ProductSearchCondition;
-import back.ecommerce.dto.response.ProductListResponse;
-import back.ecommerce.repository.ProductQueryDslRepository;
+import back.ecommerce.dto.response.product.ProductDto;
+import back.ecommerce.dto.request.product.ProductSearchCondition;
+import back.ecommerce.dto.response.product.ProductListResponse;
+import back.ecommerce.repository.product.ProductQueryDslRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
