+++ src/test/java/back/ecommerce/controller/ProductSearchControllerTest.java
@@ -15,10 +15,10 @@ import org.springframework.context.annotation.Import;
 import org.springframework.test.web.servlet.MockMvc;
 
 import back.ecommerce.domain.product.Category;
-import back.ecommerce.dto.ProductDto;
-import back.ecommerce.dto.ProductSearchCondition;
-import back.ecommerce.dto.response.ProductListResponse;
-import back.ecommerce.service.ProductService;
+import back.ecommerce.dto.response.product.ProductDto;
+import back.ecommerce.dto.request.product.ProductSearchCondition;
+import back.ecommerce.dto.response.product.ProductListResponse;
+import back.ecommerce.service.product.ProductService;
 
 @WebMvcTest(ProductSearchController.class)
 @Import(MockAuthProviderConfig.class)
