+++ src/test/java/back/ecommerce/repository/ProductRepositoryTest.java
@@ -16,9 +16,11 @@ import org.springframework.data.domain.PageRequest;
 import back.ecommerce.constant.PageConstant;
 import back.ecommerce.domain.product.Category;
 import back.ecommerce.domain.product.Product;
-import back.ecommerce.dto.ProductDto;
-import back.ecommerce.dto.ProductSearchCondition;
-import back.ecommerce.dto.ProductSortCondition;
+import back.ecommerce.dto.response.product.ProductDto;
+import back.ecommerce.dto.request.product.ProductSearchCondition;
+import back.ecommerce.dto.request.product.ProductSortCondition;
+import back.ecommerce.repository.product.ProductQueryDslRepository;
+import back.ecommerce.repository.product.ProductRepository;
 
 @DataJpaTest
 @Import(QueryDSLRepoConfig.class)
