+++ src/test/java/back/ecommerce/service/ProductServiceTest.java
@@ -15,11 +15,12 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 
 import back.ecommerce.domain.product.Category;
-import back.ecommerce.dto.ProductDto;
-import back.ecommerce.dto.ProductSearchCondition;
-import back.ecommerce.dto.ProductSortCondition;
-import back.ecommerce.dto.response.ProductListResponse;
-import back.ecommerce.repository.ProductQueryDslRepository;
+import back.ecommerce.dto.response.product.ProductDto;
+import back.ecommerce.dto.request.product.ProductSearchCondition;
+import back.ecommerce.dto.request.product.ProductSortCondition;
+import back.ecommerce.dto.response.product.ProductListResponse;
+import back.ecommerce.repository.product.ProductQueryDslRepository;
+import back.ecommerce.service.product.ProductService;
 
 @ExtendWith(MockitoExtension.class)
 class ProductServiceTest {
